release/1 admin v1.0.1

+++ client/src/providers/dataProvider.ts
@@ -1,31 +1,37 @@
 import { DataProvider, withLifecycleCallbacks } from "react-admin";
 
-import { convertToBase64 } from "../utils/convertToBase64";
-import { formatObjId } from "../utils/formatObjId";
+import { convertToBase64, formatObjId } from "../utils";
 
 import { $api } from "../api/http";
 
 import qs from "qs";
 
+const INVALID_RESOURCE_OPERATIONS = {
+	create: ["categories"],
+	delete: ["categories"],
+	deleteMany: ["categories"]
+};
+
 // @ts-ignore. don't need to use updateMany / getManyReference
 const provider: DataProvider = {
 	getOne: async (res, par) => {
 		const {
-			data: { card }
+			data: { data }
 		} = await $api.get(`/${res}/${par.id}`);
-		return { data: formatObjId(card) };
+		return { data: formatObjId(data) };
 	},
 	getMany: async (res, par) => {
 		const query = {
 			filter: JSON.stringify({ id: par.ids })
 		};
-		const url = `/${res}?${qs.stringify(query)}`;
+
+		const url = `/${res}/many?${qs.stringify(query)}`;
 
 		const {
-			data: { cards }
+			data: { data }
 		} = await $api.get(url);
 
-		const formatted = formatObjId(cards);
+		const formatted = formatObjId(data);
 
 		return { data: [...(Array.isArray(formatted) ? formatted : [])] };
 	},
@@ -36,60 +42,66 @@ const provider: DataProvider = {
 		const query = {
 			sort: JSON.stringify([field, order]),
 			range: JSON.stringify([(page - 1) * perPage, page * perPage - 1]),
-			title: JSON.stringify(par.filter.q)
+			q: JSON.stringify(par.filter.q)
 		};
 
 		const url = `/${res}?${qs.stringify(query)}`;
 
 		const {
-			data: { cards, total }
+			data: { data, total }
 		} = await $api.get(url);
 
-		const formatted = formatObjId(cards);
+		const formatted = formatObjId(data);
 
 		return { data: [...(Array.isArray(formatted) ? formatted : [])], total };
 	},
 	create: async (res, par) => {
-		const { data } = await $api.post(`/${res}`, { ...par.data });
-		return { data: formatObjId(data) };
+		if (!INVALID_RESOURCE_OPERATIONS.create.includes(res)) {
+			const {
+				data: { data }
+			} = await $api.post(`/${res}`, { ...par.data });
+			return { data: formatObjId(data) };
+		}
+
+		return Promise.reject("You can't create this resource");
 	},
 	update: async (res, par) => {
 		const {
-			data: { card }
-		} = await $api.put(`/${res}/${par.id}`, { ...par.data });
-		return { data: formatObjId(card) };
+			data: { data }
+		} = await $api.put(`/${res}/${par.id}`, {
+			...par.data,
+			previousData: par.previousData
+		});
+		return { data: formatObjId(data) };
 	},
 	delete: async (res, par) => {
-		const {
-			data: { card }
-		} = await $api.delete(`/${res}/${par.id}`);
-		return { data: formatObjId(card) };
+		if (!INVALID_RESOURCE_OPERATIONS.delete.includes(res)) {
+			const {
+				data: { data }
+			} = await $api.delete(`/${res}/${par.id}`);
+			return { data: formatObjId(data) };
+		}
+
+		return Promise.reject("You can't delete this resource");
 	},
 	deleteMany: async (res, par) => {
-		const url = `/${res}/many/${JSON.stringify(par.ids)}`;
+		if (!INVALID_RESOURCE_OPERATIONS.deleteMany.includes(res)) {
+			const url = `/${res}/many/${JSON.stringify(par.ids)}`;
 
-		const {
-			data: { ids }
-		} = await $api.delete(url);
+			const {
+				data: { ids }
+			} = await $api.delete(url);
 
-		return { data: formatObjId(ids) };
+			return { data: formatObjId(ids) };
+		}
+
+		return Promise.reject("You can't delete this resources");
 	}
 };
 
 export const dataProvider = withLifecycleCallbacks(provider, [
 	{
-		resource: "cards",
-		beforeCreate: async (params, dataProvider: DataProvider) => {
-			const image = await convertToBase64(params.data.image);
-
-			return {
-				...params,
-				data: {
-					...params.data,
-					image
-				}
-			};
-		},
+		resource: "categories",
 		beforeUpdate: async (params, dataProvider: DataProvider) => {
 			const { image, ...data } = params.data;
